tree-wide: get rid of dfuzzer.h
Everything that was left in it was used only in `dfuzzer.c`, so let's
mark all the functions as static and drop the file completely.
