+++ src/bus.c
@@ -2,7 +2,6 @@
 #include <stdlib.h>
 
 #include "bus.h"
-#include "dfuzzer.h"
 #include "log.h"
 #include "util.h"
